Backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
python-multipart
pandas
matplotlib
seaborn
scikit-learn
kagglehub
